requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
redis==5.0.1
psycopg2-binary==2.9.9
influxdb-client==1.38.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyYAML==6.0.1

# 数据科学与量化分析
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
statsmodels==0.14.0
scikit-learn==1.3.2
TA-Lib==0.4.28
pykalman==0.9.5
hmmlearn==0.3.0
arch==6.2.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
mypy==1.7.1
flake8==6.1.0
pre-commit==3.6.0

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0

# 其他工具
click==8.1.7
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
jupyter==1.0.0